# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Neoting API
  version: 0.1.0-draft
  summary: Chat-first document-to-bookkeeping platform for UK accounting practices.
  description: |
    The single API. The public API is this same API with scoped OAuth clients —
    there is no second door (D10, Governance §3).

    Money is integer pence. Nothing changes state outside the Review → Approve
    path. Tenancy is row-level security, never the URL. Every extracted value
    carries its provenance class. Errors are RFC 7807 `problem+json` with a
    stable `NT-` code.

    The reasoning behind each of those, and the conventions a consumer needs, are
    in `packages/contracts/CLAUDE.md` and in the header comment of
    `openapi.yaml`. They are deliberately not repeated here: this field is
    stamped into the header of every generated file.
  contact:
    name: Shakib Bin Kabir — tech lead
    url: https://github.com/Neovogent/neoting
  x-nt-governs:
    sourceOfTruth: docs/Source_Of_Truth.md v1.4 — §4 Stage 1-2, §8.2, §15
    governance: docs/Engineering_Governance.md v1.4 — §3, §5.2, §10, §13.4
    schema: prisma/schema.prisma
servers:
  - url: http://localhost:3000/v1
    description: |
      Local — docker compose. The canonical development environment.

      3000, not 3001. This is the port the API actually listens on and it is
      the same number in all four places that have to agree: `PORT` in
      `apps/api/src/config/env.ts`, `EXPOSE` in `apps/api/Dockerfile`,
...
security:
  - workspaceSession: []
...
paths:
...
  /auth/sessions:
    post:
      operationId: createSession
      summary: Log in
      description: |
        Verifies email + password + TOTP and issues the workspace session as an
        HttpOnly `nt_session` cookie (12 h). The response body is empty on
        purpose — the cookie IS the session, and `GET /me` is the read surface
        for who you now are.

        Server-side this changes no product state: the session is a stateless
        signed token, there is no session row, and nothing that Review →
        Approve governs happens here. That is why the side-effect class is
        `none` and no `Idempotency-Key` is required — retrying a login simply
        issues a fresh cookie.

        Every credential failure — unknown email, wrong password, wrong TOTP —
        returns the same `401 NT-AUTH-003`. Distinguishing them would hand an
        account-enumeration oracle to whoever is guessing.
      tags: [auth]
      x-nt-side-effect: none
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCreateRequest' }
      responses:
        '204':
          description: Logged in. The session travels only as the `nt_session` cookie.
          headers:
            Set-Cookie:
              description: '`nt_session=<signed token>; HttpOnly; SameSite=Lax; Path=/; Expires=<+12h>`'
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Credentials did not verify — `NT-AUTH-003`, uniformly.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: https://neoting.com/problems/NT-AUTH-003
                title: Invalid credentials
                status: 401
                code: NT-AUTH-003
                detail: The email, password or verification code did not match.
                traceId: 01J8XN2ZQ5R7TQ9E3M4K6P
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /document-uploads/{uploadId}/complete:
    parameters:
      - name: uploadId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: completeDocumentUpload
      summary: Finish an upload and enter the pipeline
      description: |
        Step two. Called after the `PUT` succeeds. Verifies the object landed,
        creates the `Document` row in state `RECEIVED`, and enqueues
        sanitisation.

        Returns immediately with the document in `RECEIVED` or `PROCESSING` —
        extraction is a queued job, never inline (Governance §7). Poll
        `GET /documents/{id}` or watch the processing log for progress; the
        per-item ETA the UI shows comes from the queue, and a manual-entry bypass
        stays available the whole time (SoT Stage 2).

        With `splitMode: AUTO_SPLIT` this returns the parent document. Children
        appear as separate documents with `parentDocumentId` set once splitting
        completes, so a caller that sees one document and no children has caught
        the job mid-flight, not lost anything.

        This is also step two for portal uploads (`POST /portal/uploads`),
        which is why the portal bearer is accepted alongside the workspace
        session — a delegated session completes the intents it created, and
        the RLS delegated policies keep it inside its own grant (Governance
        §5.2). One completion path, two trust levels, no second door.
      tags: [ingestion]
      x-nt-side-effect: ingest
      security:
        - workspaceSession: []
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [byteHash]
              properties:
                byteHash:
                  type: string
                  description: |
                    SHA-256 of the uploaded bytes, hex, lowercase. Computed
                    client-side and verified server-side. It is also the exact
                    re-send signal for deduplication (SoT Stage 6).
                  pattern: '^[a-f0-9]{64}$'
      responses:
        '201':
          description: The document is in the pipeline.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: |
            The byte hash does not match what landed in storage, or this upload
            was already completed with a different hash.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ==========================================================================
  # THE REVIEW → APPROVE SPINE
  # ==========================================================================
...
  /portal/sessions:
    post:
      operationId: createPortalSession
      summary: Open a portal session from a chase link
      description: |
        Exchanges the SMS link token plus the OTP for a portal session scoped
        to exactly the items the granting chase requested and nothing else
        (Governance §5.2). The link is deliberately forwardable to whoever
        physically holds the paperwork; the audit trail records
        requested-from separately from uploaded-by-delegated-session.

        The link token is stateless — HMAC over the chase id and expiry, the
        upload-intent pattern. Verifying it plus the OTP creates the
        `otp_sessions` row that the RLS delegated policies scope everything
        else to. Demo builds verify the OTP against a fixed code
        (`OTP_MODE=demo`); Twilio Verify replaces the verifier, not this
        shape.

        Every verification failure — unknown token, expired link, wrong OTP —
        returns the same `401 NT-OTP-001`. Distinguishing them would tell a
        guesser which links exist (the NT-AUTH-003 stance, applied here).
      tags: [portal]
      x-nt-side-effect: ingest
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortalSessionCreateRequest' }
      responses:
        '201':
          description: The scoped session. Send `token` as a bearer on every portal call.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Link token or OTP failed verification — `NT-OTP-001`, uniformly.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: https://neoting.com/problems/NT-OTP-001
                title: Verification failed
                status: 401
                code: NT-OTP-001
                detail: The link or verification code did not verify. Request a fresh link if this one has expired.
                traceId: 01J8XN2ZQ5R7TQ9E3M4K6P
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /portal/context:
    get:
      operationId: getPortalContext
      summary: What this portal session may see
      description: |
        The requested items — supplier, amount, date — and the whole of what
        the session may see. The RLS delegated-session policies enforce the
        boundary below the handler; this endpoint could not widen it if it
        tried, which is the point (Governance §5.2).
      tags: [portal]
      x-nt-side-effect: none
      security:
        - portalSession: []
      responses:
        '200':
          description: The chased items this session exists to collect.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalContext' }
        '401':
          description: Portal session missing, invalid or expired — `NT-OTP-002`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /portal/uploads:
    post:
      operationId: createPortalUpload
      summary: Start a delegated upload
      description: |
        The portal's step one of two — the same intent → presigned `PUT` →
        complete flow as `/document-uploads`, under the delegated scope. The
        API never receives the bytes; that is what keeps the lightest surface
        in the product light on a bad connection in a car park.

        Complete it with `POST /document-uploads/{uploadId}/complete`, which
        accepts the portal bearer for exactly this reason. The document then
        runs the full pipeline; extraction is compared against the chased
        items, a match auto-closes the chase (SoT Stage 8.5), and a mismatch
        is fed back in the portal naming the difference — never silently
        accepted, never silently dropped.
      tags: [portal]
      x-nt-side-effect: ingest
      security:
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortalUploadRequest' }
      responses:
        '201':
          description: Upload intent created. `PUT` the bytes to `upload.url`.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentUpload' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Portal session missing, invalid or expired — `NT-OTP-002`.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '413':
          description: The declared size exceeds the 25 MB client-channel cap. Client-side compression should have prevented this for camera photos.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '415':
          description: The declared MIME type is not on the allowlist for this channel.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # --------------------------------------------------------------------------
  # PORTAL ONBOARDING — a session that is NOT the consequence of a chase
  # (ID LAW batch, D47 / SoT §24.5)
  #
  # `POST /portal/sessions` above can only exist downstream of a chase: it takes
  # a chase link token, resolves the chase, and 401s if it is gone. But an
  # invited client has no chase yet — nobody has asked them for anything, they
  # are being asked to set themselves up. `OtpSessionScope.ONBOARDING` was
  # declared in the schema for exactly this and nothing implemented it, because
  # `otp_sessions.business_id` was NOT NULL and a pre-client session had nowhere
  # to live. The ID LAW batch made it nullable; these two operations are the
  # surface that uses it.
  #
  # Two steps, not one, and the split is not ceremony: the code has to be
  # DELIVERED between them. Step one sends it, step two spends it.
  # --------------------------------------------------------------------------
  /portal/sign-in-codes:
    post:
      operationId: createPortalSignInCode
      summary: Email a six-digit sign-in code
      description: |
        The client follows the setup link their accountant sent, gives the email
        address they were registered with, and receives a six-digit code.

        **Always `202`, whatever happened.** Whether that address is registered
        on that workspace is not something an unauthenticated caller may learn —
        the `NT-AUTH-003` stance, applied here. The email is the channel that
        distinguishes the two outcomes, and it goes to the address, not to the
        caller.

        **The code is a credential.** It is never logged, never placed in a URL,
        never returned in a response or an error, and not in development either.
        Rate-limited per address AND per IP: a six-digit code with unlimited
        attempts is a four-digit code by lunchtime.

        There is no SMS here. D47 names the client-registration route as a short
        setup link plus the company general email, and the copy on every screen
        must say "emailed", never "texted".
      tags: [portal]
      x-nt-side-effect: ingest
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortalSignInCodeRequest' }
      responses:
        '202':
          description: |
            Accepted. If that address is registered on this workspace, a code is
            on its way. Empty by design — see the description.
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }
  /portal/onboarding-sessions:
    post:
      operationId: createPortalOnboardingSession
      summary: Open a portal session from a setup link
      description: |
        Exchanges the setup link, the registered address and the emailed code
        for a portal session — the same `PortalSession` bearer the chase path
        issues, opened under `OtpSessionScope.ONBOARDING` rather than
        `DELEGATED_UPLOAD`.

        **An onboarding session grants no items.** A delegated session's whole
        authority is its list of chased item ids; this one has no chase and
        therefore no list, which is precisely why it needed a scope of its own
        rather than an empty grant. An empty grant reads to a human as "no
        restriction" and denies everything in SQL — a gap between what a
        reviewer sees and what the database does is the shape of bug this
        contract exists to prevent.

        Every verification failure — unknown link, expired link, wrong address,
        wrong code — returns the same `401 NT-OTP-001`, for the reason the chase
        path gives.
      tags: [portal]
      x-nt-side-effect: ingest
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PortalOnboardingSessionRequest' }
      responses:
        '201':
          description: The session. Send `token` as a bearer on every portal call.
          headers:
            NT-Trace-Id: { $ref: '#/components/headers/TraceId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortalSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: Link, address or code failed verification — `NT-OTP-001`, uniformly.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409': { $ref: '#/components/responses/IdempotencyConflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ==========================================================================
  # PUBLISHING — read surface (METH Stage 2, issue #120)
  # ==========================================================================
...
components:
  securitySchemes:
    workspaceSession:
      type: apiKey
      in: cookie
      name: nt_session
      description: |
        The accountant workspace session. Device-bound, rotating refresh token,
        absolute lifetime ≤ 30 days, idle timeout ≤ 24 h (Governance §11.1).
        Every state-changing browser request is additionally CSRF-protected.

    portalSession:
      type: http
      scheme: bearer
      description: |
        A client's OTP portal session. Scoped to exactly the items the granting
        chase requested and nothing else — enforced by RLS via
        `app.session_scope = 'delegated_upload'` plus the granted item IDs
        (Governance §5.2), not by the handler.

        The link and OTP are deliberately forwardable to whoever physically holds
        the document; the audit trail records *requested-from* separately from
        *uploaded-by-delegated-session*.

    oauth2:
      type: oauth2
      description: |
        Public API clients (D10, 185-vote gap). The same API, scoped — there is
        no second door. `documents:write` grants proposal creation and never
        approval: approving is a human UI action by definition.
      flows:
        authorizationCode:
          authorizationUrl: https://api.neoting.neovogent.com/v1/oauth/authorize
          tokenUrl: https://api.neoting.neovogent.com/v1/oauth/token
          scopes:
            documents:read: Read documents, extractions and processing logs
            documents:write: Create documents and propose document actions
            proposals:read: Read action proposals

    stripeSignature:
      type: apiKey
      in: header
...
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated UUIDv4, unique per logical operation. Replays return the
        original result rather than acting twice. Reusing a key with a different
        payload is `409` — that is a client bug, and failing loudly beats
        silently doing the wrong thing to someone's books.
      schema: { type: string, format: uuid }
  headers:
    TraceId:
      description: |
        The trace id born at this entry point and carried through every service,
        job, model call and adapter call it causes. Quote it in a bug report and
        the whole journey is one Journey Inspector lookup.
      schema: { type: string }
...
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: { type: integer }
  responses:
    BadRequest:
      description: The request failed schema validation at the boundary.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://neoting.com/problems/NT-VAL-001
            title: Validation failed
            status: 400
            code: NT-VAL-001
            detail: One or more fields are invalid.
            traceId: 01J8XN2ZQ5R7TQ9E3M4K6P
            errors:
              - field: byteSize
                message: must be a positive integer
    Unauthenticated:
      description: No valid session or token.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    NotFound:
      description: |
        No such record — or one the caller's RLS context does not reach. The two
        are deliberately indistinguishable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    IdempotencyConflict:
      description: This `Idempotency-Key` was already used with a different payload.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    RateLimited:
      description: Rate limit exceeded (Governance §11.8).
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: |
        Unexpected. The response carries the trace id and nothing else — internal
        detail and PII never reach a client (Governance §3).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Problem:
      type: object
      description: RFC 7807 problem detail. The only error shape this API returns.
      required: [type, title, status, code, traceId]
      properties:
        type:
          type: string
          format: uri
          description: |
            Points at the runbook page for `code`. Governance §13.4 requires
            every code to have one before it passes review, so this link is never
            dead.
          examples: ['https://neoting.com/problems/NT-ING-004']
        title: { type: string, description: Short, human, stable for the code. }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: What went wrong for this occurrence, in plain English, never blaming the user.
        traceId: { type: string }
        errors:
          type: array
          description: Field-level detail on validation failures.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    SessionCreateRequest:
      type: object
      required: [email, password, totp]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        totp:
          type: string
          pattern: '^[0-9]{6}$'
          description: |
            Six digits. Demo builds verify against a fixed code
            (`OTP_MODE=demo`); real TOTP replaces the verifier, not this shape.

    # ---- the signup chain (launch stage A14, issue #195) ------------------
...
    PortalSessionCreateRequest:
      type: object
      required: [linkToken, otp]
      properties:
        linkToken:
          type: string
          description: From the SMS link. Stateless — HMAC over the chase id and expiry.
        otp:
          type: string
          pattern: '^[0-9]{6}$'
          description: |
            Six digits. Demo builds verify against a fixed code
            (`OTP_MODE=demo`); real verification replaces the verifier, not
            this shape.
...
    PortalOnboardingSessionRequest:
      type: object
      required: [email, otp]
      properties:
        setupToken:
          type: string
          description: |
            Optional, and must match how the code was requested — see
            `PortalSignInCodeRequest.setupToken`.
        email: { type: string, format: email }
        otp:
          type: string
          pattern: '^[0-9]{6}$'
          description: |
            The six digits from the email. Single use, short expiry, attempt
            limited. Never logged, never in a URL, never in a response.

    # ---- publishing (SoT Stage 10 — METH Stage 2, issue #120) ---------------
...